Implement the whole framework for the experiment (using glutMainLoop())

+++ code/experiment/CylinderObject.cpp
@@ -48,7 +48,7 @@ void CylinderObject::setRandPara()
     this->radius = this->radiusRange[randIndex];
 }
 
-BOOL CylinderObject::adjustAsptRatio(int delta)
+BOOL CylinderObject::adjustAsptRatio(GLfloat delta)
 {
     if(adjZAsptRatio + delta <= 0)
     {
